Removed `check_convergence` function from the controller and replaced with similiarly convenient function
