+++ pySDC/implementations/convergence_controller_classes/check_convergence.py
@@ -26,6 +26,28 @@ class CheckConvergence(ConvergenceController):
         """
         return {"control_order": +200, **params}
 
+    @staticmethod
+    def check_convergence(S):
+        """
+        Check the convergence of a single step.
+
+        Args:
+            S (pySDC.Step): The current step
+
+        Returns:
+            bool: Convergence status of the step
+        """
+        # do all this on the finest level
+        L = S.levels[0]
+        L.sweep.compute_residual()
+
+        # get residual and check against prescribed tolerance (plus check number of iterations
+        res = L.status.residual
+        converged = S.status.iter >= S.params.maxiter or res <= L.params.restol or S.status.force_done
+        if converged is None:
+            converged = False
+        return converged
+
     def check_iteration_status(self, controller, S, **kwargs):
         """
         Routine to determine whether to stop iterating (currently testing the residual + the max. number of iterations)
@@ -37,16 +59,7 @@ class CheckConvergence(ConvergenceController):
         Returns:
             None
         """
-
-        # do all this on the finest level
-        L = S.levels[0]
-        L.sweep.compute_residual()
-
-        # get residual and check against prescribed tolerance (plus check number of iterations
-        res = L.status.residual
-        converged = S.status.iter >= S.params.maxiter or res <= L.params.restol or S.status.force_done
-        if converged is not None:
-            S.status.done = converged
+        S.status.done = self.check_convergence(S)
 
         if "comm" in kwargs.keys():
             self.communicate_convergence(controller, S, **kwargs)
